move method to helper
